+++ lambda/custom/handlers/cancel-stop-intent-handler.js
@@ -1,0 +1,17 @@
+const CancelAndStopIntentHandler = {
+    canHandle(handlerInput) {
+      return handlerInput.requestEnvelope.request.type === 'IntentRequest'
+        && (handlerInput.requestEnvelope.request.intent.name === 'AMAZON.CancelIntent'
+          || handlerInput.requestEnvelope.request.intent.name === 'AMAZON.StopIntent');
+    },
+    handle(handlerInput) {
+      const speechText = 'Goodbye!';
+  
+      return handlerInput.responseBuilder
+        .speak(speechText)
+        .withSimpleCard('Hello World', speechText)
+        .getResponse();
+    },
+  };
+
+  exports.CancelAndStopIntentHandler = CancelAndStopIntentHandler;
