+++ lambda/custom/handlers/error-handler.js
@@ -1,0 +1,15 @@
+const ErrorHandler = {
+    canHandle() {
+      return true;
+    },
+    handle(handlerInput, error) {
+      console.log(`Error handled: ${error.message}`);
+  
+      return handlerInput.responseBuilder
+        .speak('Sorry, I can\'t understand the command. Please say again.')
+        .reprompt('Sorry, I can\'t understand the command. Please say again.')
+        .getResponse();
+    },
+  };
+
+  exports.ErrorHandler = ErrorHandler
