+++ lambda/custom/handlers/help-intent-handler.js
@@ -1,0 +1,17 @@
+const HelpIntentHandler = {
+    canHandle(handlerInput) {
+      return handlerInput.requestEnvelope.request.type === 'IntentRequest'
+        && handlerInput.requestEnvelope.request.intent.name === 'AMAZON.HelpIntent';
+    },
+    handle(handlerInput) {
+      const speechText = 'You can say hello to me!';
+  
+      return handlerInput.responseBuilder
+        .speak(speechText)
+        .reprompt(speechText)
+        .withSimpleCard('Hello World', speechText)
+        .getResponse();
+    },
+  };
+
+  exports.HelpIntentHandler = HelpIntentHandler;
