+++ lambda/custom/handlers/launch-request-handler.js
@@ -1,5 +1,3 @@
-const Alexa = require('ask-sdk-core');
-
 const LaunchRequestHandler = {
     canHandle(handlerInput) {
       return handlerInput.requestEnvelope.request.type === 'LaunchRequest';
