+++ lambda/custom/index.js
@@ -5,6 +5,8 @@ const Alexa = require('ask-sdk-core');
 
 const LaunchRequestHandler = require('./handlers/launch-request-handler');
 const SessionEndedRequestHandler = require('./handlers/session-ended-handler');
+const CancelAndStopIntentHandler = require('./handlers/cancel-stop-intent-handler');
+const HelpIntentHandler = require('./handlers/help-intent-handler');
 
 const HelloWorldIntentHandler = {
   canHandle(handlerInput) {
@@ -21,60 +23,14 @@ const HelloWorldIntentHandler = {
   },
 };
 
-const HelpIntentHandler = {
-  canHandle(handlerInput) {
-    return handlerInput.requestEnvelope.request.type === 'IntentRequest'
-      && handlerInput.requestEnvelope.request.intent.name === 'AMAZON.HelpIntent';
-  },
-  handle(handlerInput) {
-    const speechText = 'You can say hello to me!';
-
-    return handlerInput.responseBuilder
-      .speak(speechText)
-      .reprompt(speechText)
-      .withSimpleCard('Hello World', speechText)
-      .getResponse();
-  },
-};
-
-const CancelAndStopIntentHandler = {
-  canHandle(handlerInput) {
-    return handlerInput.requestEnvelope.request.type === 'IntentRequest'
-      && (handlerInput.requestEnvelope.request.intent.name === 'AMAZON.CancelIntent'
-        || handlerInput.requestEnvelope.request.intent.name === 'AMAZON.StopIntent');
-  },
-  handle(handlerInput) {
-    const speechText = 'Goodbye!';
-
-    return handlerInput.responseBuilder
-      .speak(speechText)
-      .withSimpleCard('Hello World', speechText)
-      .getResponse();
-  },
-};
-
-const ErrorHandler = {
-  canHandle() {
-    return true;
-  },
-  handle(handlerInput, error) {
-    console.log(`Error handled: ${error.message}`);
-
-    return handlerInput.responseBuilder
-      .speak('Sorry, I can\'t understand the command. Please say again.')
-      .reprompt('Sorry, I can\'t understand the command. Please say again.')
-      .getResponse();
-  },
-};
-
 const skillBuilder = Alexa.SkillBuilders.custom();
 
 exports.handler = skillBuilder
   .addRequestHandlers(
     LaunchRequestHandler.LaunchRequestHandler,
     HelloWorldIntentHandler,
-    HelpIntentHandler,
-    CancelAndStopIntentHandler,
+    HelpIntentHandler.HelpIntentHandler,
+    CancelAndStopIntentHandler.CancelAndStopIntentHandler,
     SessionEndedRequestHandler.SessionEndedRequestHandler
   )
   .addErrorHandlers(ErrorHandler)
